fix(ci): enable clang-tidy bugprone-signed-char-misuse / cert-str34-c
- commit2: added explicit cast to relevant fcn calls where assigned var is int8_t
- commit3: added buffer length guard to ReadOrdIDs state

+++ src/lib/sensor_calibration/Gyroscope.cpp
@@ -145,7 +145,7 @@ void Gyroscope::ParametersUpdate()
 		return;
 	}
 
-	_calibration_index = FindCurrentCalibrationIndex(SensorString(), _device_id);
+	_calibration_index = static_cast<int8_t>(FindCurrentCalibrationIndex(SensorString(), _device_id));
 
 	if (_calibration_index == -1) {
 		// no saved calibration available
