+++ src/modules/sensors/voted_sensors_update.cpp
@@ -88,7 +88,7 @@ void VotedSensorsUpdate::parametersUpdate()
 		    && (imu.get().accel_device_id != 0) && (imu.get().gyro_device_id != 0)) {
 
 			// find corresponding configured accel priority
-			int8_t accel_cal_index = calibration::FindCurrentCalibrationIndex("ACC", imu.get().accel_device_id);
+			int8_t accel_cal_index = static_cast<int8_t>(calibration::FindCurrentCalibrationIndex("ACC", imu.get().accel_device_id));
 
 			if (accel_cal_index >= 0) {
 				// found matching CAL_ACCx_PRIO
@@ -111,7 +111,7 @@ void VotedSensorsUpdate::parametersUpdate()
 			}
 
 			// find corresponding configured gyro priority
-			int8_t gyro_cal_index = calibration::FindCurrentCalibrationIndex("GYRO", imu.get().gyro_device_id);
+			int8_t gyro_cal_index = static_cast<int8_t>(calibration::FindCurrentCalibrationIndex("GYRO", imu.get().gyro_device_id));
 
 			if (gyro_cal_index >= 0) {
 				// found matching CAL_GYROx_PRIO
